requirements-test.txt
-r requirements.txt
pytest==6.2.4
pytest-django==4.3.0
model-bakery==1.2.1
